+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 
 * [X] 更新版本号。
 * [X] breeze-data 补充注释，搁置测试。
+* [X] breeze-core 移除Range的中缀扩展，因为可能返回`ClosedRange<T>`或`Pair`。
 * [ ] DELAY breeze-core Uri构建和编码的扩展。
 * [ ] DELAY breeze-core 利用反射的类型转换的扩展。
 * [ ] DELAY breeze-text en 单数与复数的转化。